+++ src/SetHandler.cpp
@@ -89,7 +89,7 @@ void SetHandler::normalizeDOCFeatures(Normalizer* pNorm) {
 
 int const SetHandler::getLabel(int setPos) {
   assert(setPos >= 0 && setPos < (signed int)subsets_.size());
-  return subsets_[setPos]->getLabel();
+  return subsets_[static_cast<std::size_t>(setPos)]->getLabel();
 }
 
 int SetHandler::readTab(istream& dataStream, SanityCheck*& pCheck) {
@@ -149,7 +149,7 @@ bool SetHandler::isDefaultDirectionLine(const std::string& defaultDirectionLine)
 
 int SetHandler::getNumFeatures(const std::string& line, int optionalFieldCount) {
   TabReader reader(line);
-  reader.skip(2u + optionalFieldCount); // remove id, label and optional fields
+  reader.skip(static_cast<std::size_t>(2 + optionalFieldCount)); // remove id, label and optional fields
   
   reader.readDouble();
   int numFeatures = 0;
@@ -165,7 +165,7 @@ void SetHandler::getFeatureNames(const std::string& headerLine,
     int numFeatures, int optionalFieldCount, FeatureNames& featureNames) {
   TabReader reader(headerLine);
   // removes enumerator, label and if present optional fields
-  reader.skip(2u + optionalFieldCount);
+  reader.skip(static_cast<std::size_t>(2 + optionalFieldCount));
   int numFeatLeft = numFeatures;
   while (!reader.error()) {
     std::string tmp = reader.readString();
@@ -182,7 +182,7 @@ bool SetHandler::getInitValues(const std::string& defaultDirectionLine,
     int optionalFieldCount, std::vector<double>& init_values) {
   TabReader reader(defaultDirectionLine);
   // removes enumerator, label and if present optional fields
-  reader.skip(2u + optionalFieldCount);
+  reader.skip(static_cast<std::size_t>(2 + optionalFieldCount));
   
   bool hasDefaultValues = false;
   unsigned int ix = 0;
@@ -419,7 +419,7 @@ ScanId SetHandler::getScanId(const std::string& psmLine, int& label,
       }
     }
   }
-  if (!hasScannr) scanId.first = lineNr;
+  if (!hasScannr) scanId.first = static_cast<int>(lineNr);
   
   return scanId;
 }
@@ -486,7 +486,7 @@ int SetHandler::readAndScoreTab(istream& dataStream,
     oss << "ERROR: Reading tab file, too many default values present." << std::endl;
     if (NO_TERMINATE) {
       cerr << oss.str() << "No-terminate flag set: ignoring error and trimming default value vector." << std::endl;
-      init_values.resize(numFeatures);
+      init_values.resize(static_cast<std::size_t>(numFeatures));
     } else {
       throw MyException(oss.str());
     }
